Continued interaction rework
Converted interaction to use a specific global class component rather than duck typing. This serves to help guarantee an API for interaction without needing to check  against each individual property and function. Because that's what classes are for. Why in the hell isn't this just a thing in Godot? Give me interfaces you cowards! Or at least an actual Object Oriented structure